api-platform2/deployments/docker/requirements.txt
Flask
Flask-BabelEx==0.9.3
Flask-Images==2.1.2
Flask-Login==0.4.1
Flask-Migrate==2.1.1
Flask-MySQL==1.4.0
Flask-Principal==0.4.0
Flask-RESTful==0.3.5
Flask-Security==3.0.0
Flask-SocketIO==3.0.1
Flask-SQLAlchemy==2.3.2
Flask-HTTPAuth==3.3.0
uuid
jwt
Flask-JWT
kubernetes==9.0.0
pyyaml
mysqlclient
